[Fixes] Abstract functions in test_url.py to util.py
This abstraction is carried out so as to reuse those functions in test_user.py
